app/api/requirements.txt
# Rythmiq One FastAPI Gateway - Phase 2A
# Python dependencies

fastapi>=0.109.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.30.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
PyJWT>=2.8.0,<3.0.0
supabase>=2.3.0,<3.0.0
boto3>=1.34.0,<2.0.0
httpx>=0.26.0,<0.28.0
